requirements.txt
# Google Agent Development Kit
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.38.0

# Crawlee Web Scraping
crawlee[playwright]>=0.0.7

# Web Scraping & HTTP
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0

# Email Validation
email-validator>=2.1.0
dnspython>=2.4.0

# Data Processing
pandas>=2.0.0
python-dateutil>=2.8.0

# Configuration & Environment
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
aiohttp>=3.9.0
asyncio>=3.4.3
tldextract>=5.0.0

# CLI
click>=8.1.0
rich>=13.0.0

# Logging
loguru>=0.7.0
